+++ src/constants/vsCommands.constants.ts
@@ -7,6 +7,8 @@ export const vsCommands = {
 	buildFolder: "autokitteh.buildFolder",
 	openBaseURLInputDialog: "autokitteh.openBaseURLInputDialog",
 	openConfigSetupWalkthrough: "autokitteh.openConfigSetupWalkthrough",
+	buildProject: "autokitteh.buildProject",
+	runProject: "autokitteh.runProject",
 	openWalkthrough: "workbench.action.openWalkthrough",
 	baseURLUpdated: "autokitteh.baseURLUpdated",
 	showErrorMessage: "autokitteh.showErrorMessage",