+++ src/types/interfaces/index.ts
@@ -1,4 +1,2 @@
 export type { SectionRowsRange } from "@type/interfaces/callback.interface.type";
 export type { Callback } from "@type/interfaces/callback.interface.type";
-export type { ProjectCB } from "@type/interfaces/callback.interface.type";
-export type { CallbackWStringIdParam } from "@type/interfaces/callback.interface.type";